Inclusão de model de validação de calculo de juros. Desenvolvimento de chamada de calculo de juros.

+++ DesafioTecnico/DesafioTecnico.Calculos.Dominio/Interfaces/IServicoCalculoJuros.cs
@@ -2,6 +2,6 @@
 {
     public interface IServicoCalculoJuros
     {
-        decimal CalcularJuros(double valorInicial, int meses);
+        decimal CalcularJuros(decimal valorInicial, int meses);
     }
 }
